Move dict_merge to dedicated
