+++ src/ansible_navigator/utils/functions.py
@@ -16,7 +16,6 @@ from dataclasses import dataclass
 from enum import Enum
 from pathlib import Path
 from typing import Any
-from typing import Dict
 from typing import Iterable
 from typing import List
 from typing import Mapping
@@ -158,57 +157,6 @@ def clear_screen() -> None:
     if os.environ.get("TERM_PROGRAM") in affected_terminals:
         for _line in range(shutil.get_terminal_size().lines):
             print()
-
-
-class DictMergeError(Exception):
-    """Custom exception for a dict merge error."""
-
-
-Mergeable = Optional[
-    Union[
-        bool,
-        Dict,
-        List,
-        str,
-    ]
-]
-
-
-def dict_merge_in_place_list_replace(left: Mergeable, right: Mergeable):
-    """Merges right into left in place and returns merged result
-
-    If left is a primitive, right will be used
-    If left key is a list or tuple right will be used.
-    If left and right have a common key, they will be dict merged
-    If left is missing a key from right, the key from right will be added
-
-    :param left: Dict to be merged into
-    :param right: Dict to be merged from
-    :return: Merged dict
-    """
-    key = None
-    try:
-        if left is None or isinstance(left, (str, int, float, bool)):
-            # Border case for first run or if a is a primitive
-            left = right
-        elif isinstance(left, (tuple, list)):
-            left = right
-        elif isinstance(left, dict):
-            if isinstance(right, dict):
-                for key in right:
-                    if key in left:
-                        left[key] = dict_merge_in_place_list_replace(left[key], right[key])
-                    else:
-                        left[key] = right[key]
-            else:
-                raise DictMergeError(f"Cannot merge non-dict '{right}' into dict '{left}'")
-        else:
-            raise DictMergeError(f"Not implemented '{right}' into '{left}'")
-    except TypeError as exc:
-        raise DictMergeError(
-            f"TypeError '{exc}' in key '{key}' when merging '{right}' into '{left}'",
-        ) from exc
-    return left
 
 
 # TODO: Replace this with something type-safe.
